Add screen components to extract and modulorzie the code]
