--- src/types/types.ts
@@ -1,6 +1,0 @@
-
-export interface IUserData{
-    userName : string;
-    password : string;
-    role : 'doctor' | 'patient';
-}
